simplify provider
